+++ src/ts/index.ts
@@ -10,6 +10,7 @@ export {
     FlatRecord,
     type DistanceUnit,
     type NearestFeaturesOptions,
+    type PreloadOptions,
     type FlatRecordInspect,
     type FlatRecordBlockInfo,
 } from './flat-record.js';